cleaning up and renaming things
